feat(plugins): web_fetch/run_command 迁出为独立插件 + 彻底删除 LocalToolExecutor
把 core LocalToolExecutor 剩余的 web_fetch 和 run_command 完全迁出为独立插件，
core 不再直接执行任何工具。CLI/Server 注册 fetch/command 插件（基础能力），
GUI 继续用 browser/terminal 插件（增强能力），terminal 插件额外接管 run_command。

新增 tiangong-plugin-fetch（web_fetch）：
- reqwest::blocking 获取 URL，含 SSRF 防护（拒绝本机/非公网地址）、HTML 正文
  提取、download 模式 SHA256 落盘
- 参数命名化（直接读 call.arguments JSON），thread::scope 隔离 blocking runtime
- trust_mode 仅影响 download 写路径
- 4 个测试（缺 url / localhost 拒绝 / 私网 IP 拒绝 / 未知工具）

新增 tiangong-plugin-command（run_command / run_shell / run_bash）：
- tokio::process::Command 子进程执行（CLI/Server 无 PTY）
- 完整保留命令白名单（is_allowed_command）/ 路径越界校验 /
  shell 脚本校验（复用 core common.rs）
- trust_mode=FullTrust 时跳过校验；register 时从 engine 获取 runtime_env
- dispatch_sync 同步解析校验后返回 owned Future（满足 ToolOverrideHandler 'static）
- 4 个测试（echo 执行 / 非白名单拒绝 / shell 执行 / 未知工具）

core 清理（彻底删除 LocalToolExecutor）：
- 删除 tool/web_fetch.rs、tool/run_command.rs
- tool.rs 仅保留 ToolResult / ToolExecutionRecord，删除 ToolName / ToolCall /
  ToolExecutor trait / LocalToolExecutor 结构
- runtime_env 收集逻辑迁入新模块 runtime_env.rs，RuntimeEngine 直接持有
  runtime_env 字段（替代原 tool_executor），spawn_task 改读 engine.runtime_env()
- 删除 runtime.rs 的本地工具 fallback 分支（所有工具均由插件 override 处理）
- 删除 build_tool_call_from_function（命名参数直读，废弃位置参数转换）
- basic_file_function_tools 仅保留 mark_step_completed / plugin_injection，
  删除 web_fetch / run_command spec（顺带修复 GUI 下 web_fetch spec 重复 bug）
- 移除 core 不再使用的 scraper / sha2 依赖；process 模块提升为 pub
- set_terminal_provider 移除 tool_executor 同步更新逻辑

terminal 插件接管 run_command（GUI）：
- tool_specs 新增 run_command spec
- handle 新增 run_command 分支：命令白名单 / 路径校验后经 PTY 执行
  （cwd 包装 cd && cmd），不再依赖 core LocalToolExecutor 的内置 PTY 路径

入口注册：
- CLI / Server：注册 fs + fetch + command + scheduler
- GUI：保持 browser + terminal + scheduler + fs（不注册 fetch/command）

全 workspace cargo check 通过，fs/fetch/command/scheduler 测试全绿（25 项）。
