+++ Cargo.toml
@@ -14,6 +14,8 @@ members = [
     "crates/tiangong-config",
     "crates/tiangong-scheduler",
     "crates/plugins/tiangong-plugin-browser",
+    "crates/plugins/tiangong-plugin-command",
+    "crates/plugins/tiangong-plugin-fetch",
     "crates/plugins/tiangong-plugin-fs",
     "crates/plugins/tiangong-plugin-scheduler",
     "crates/plugins/tiangong-plugin-terminal",
@@ -34,6 +36,8 @@ tiangong-llm = { path = "crates/tiangong-llm" }
 tiangong-media = { path = "crates/tiangong-media" }
 tiangong-memory = { path = "crates/tiangong-memory" }
 tiangong-plugin-browser = { path = "crates/plugins/tiangong-plugin-browser" }
+tiangong-plugin-command = { path = "crates/plugins/tiangong-plugin-command" }
+tiangong-plugin-fetch = { path = "crates/plugins/tiangong-plugin-fetch" }
 tiangong-plugin-fs = { path = "crates/plugins/tiangong-plugin-fs" }
 tiangong-plugin-scheduler = { path = "crates/plugins/tiangong-plugin-scheduler" }
 tiangong-plugin-terminal = { path = "crates/plugins/tiangong-plugin-terminal" }
