+++ crates/plugins/tiangong-plugin-command/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "tiangong-plugin-command"
+version.workspace = true
+edition = "2024"
+license = "Apache-2.0"
+
+[dependencies]
+tiangong-core = { workspace = true }
+anyhow = { workspace = true }
+serde_json = { workspace = true }
+tokio = { workspace = true, features = ["process", "time", "rt"] }
+
+[dev-dependencies]
+tempfile = { workspace = true }
+tokio = { workspace = true, features = ["macros", "rt"] }
