+++ crates/plugins/tiangong-plugin-fetch/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "tiangong-plugin-fetch"
+version.workspace = true
+edition = "2024"
+license = "Apache-2.0"
+
+[dependencies]
+tiangong-core = { workspace = true }
+reqwest = { workspace = true, default-features = false, features = ["blocking", "rustls"] }
+scraper = { workspace = true }
+sha2 = { workspace = true }
+scru128 = { workspace = true }
+anyhow = { workspace = true }
+serde = { workspace = true, features = ["derive"] }
+serde_json = { workspace = true }
+tokio = { workspace = true, features = ["rt"] }
+
+[dev-dependencies]
+tempfile = { workspace = true }
+tokio = { workspace = true, features = ["macros", "rt"] }
