+++ crates/tiangong-core/Cargo.toml
@@ -42,8 +42,6 @@ reqwest = { workspace = true, default-features = false, features = [
 ] }
 futures-util = { workspace = true }
 http = { workspace = true }
-scraper = { workspace = true }
-sha2 = { workspace = true }
 toml = { workspace = true }
 tracing = { workspace = true }
 zip = { workspace = true, default-features = false, features = ["deflate"] }
