+++ crates/tiangong-core/src/lib.rs
@@ -22,10 +22,11 @@ pub mod observe;
 pub mod permission;
 pub mod planner;
 pub mod plugin;
-mod process;
+pub mod process;
 pub mod prompt;
 pub mod react;
 pub mod runtime;
+pub mod runtime_env;
 pub mod session;
 pub mod skill;
 mod stream_throttle;
