+++ crates/tiangong-server/Cargo.toml
@@ -6,6 +6,8 @@ license = "Apache-2.0"
 
 [dependencies]
 tiangong-core = { workspace = true }
+tiangong-plugin-command = { workspace = true }
+tiangong-plugin-fetch = { workspace = true }
 tiangong-plugin-fs = { workspace = true }
 tiangong-plugin-scheduler = { workspace = true }
 tiangong-config = { workspace = true }
